Fix paddle collision and movement
The paddle and ball collision is broken because it is currently checking whether the ball is at exactly at the end of the game window. Due to the velocity of the ball, the ball is never positioned exactly at this value. Instead simply check that the ball is below the paddle. In addition the paddle is able to move beyond the game width. This movement should be restricted to within the scope of the game.